added a couple more fields to Card for spaced repetition

+++ src/main/model/Card.java
@@ -8,8 +8,8 @@ public class Card {
     private String front;            // Content on the "front" of the card
     private String back;             // Content on the "back" of the card
     private ArrayList<String> tags;  // List of tags applied to card
-    private int lastReviewDate;
-    private int reviewDate;
+    private long lastReviewDate;
+    private long reviewDate;
     private double ease;
     private int status;
